src/test/java/com/leetcode/algorithm/easy/strobogrammaticnumber/SolutionTest.java
package com.leetcode.algorithm.easy.strobogrammaticnumber;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Test;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    assertEquals(true, solution.isStrobogrammatic("69"));
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    assertEquals(true, solution.isStrobogrammatic("88"));
  }

  @Test
  public void testCase3() {
    Solution solution = new Solution();
    assertEquals(false, solution.isStrobogrammatic("962"));
  }
}
